edit posts

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -90,8 +90,19 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
+        
+            // if ($category->posts()->count()) {
+            //     return back()->withErrors(['error' => 'Cannot delete, category has posts.']);
+            // }
 
+        if ($category->posts()->count()) 
+        {
+            return back()->withErrors(['error' => 'Cannot delete, category has posts.']);
+        } else 
+        {
+            $category->delete();
+        }
+        
         return redirect()->route('categories.index');
 
     }
